test: split config flow tests by flow area
